Added collision (and changed scene accordingly) + modified player layer + modified collider offset
Added collision with tilemap and composite collider.
For now collider are on the different tilemap (wall, water, ...) though I think in the end I will just create a collider tilemap to make things way easier (cause in that case I only need a -.125 offset on the Y axis and not different offsets on different tilemaps).

Concerning colliders, here's what has changed :
- collider need to be offset by -.125 on the Y axis to actually fit the grid properly, and obstacles like water need to be offset by .5 on the Y axis (so in fine a .375 offset) so that the collision is with the "top" of the tile and not the "bottom".
- tile collider type set to "grid" from "tile"
- once again halved value for the collision check, so the check happen with the border of the tile (where the collider actually is) and not with the center of the tile (wich allowed the player to go through obastcle) [It's a difference due to the change from normal to isometric grid]

Also moved the player layer to above the wall layer, so that he would correctly be displayed behind walls.  (and tweaked the rotation on Y axis from 0 to 25 so the player sprite would better fit the isometric grid).